// src/generated/models/ApprovalSubmissionsModel.ts
export interface NSCValue {
  "@odata.type": string;
  Id: number;
  Value: string;
}

export interface FunctionValue {
  "@odata.type": string;
  Id: number;
  Value: string;
}

export interface CategoryValue {
  "@odata.type": string;
  Id: number;
  Value: string;
}

export interface SubcategoryValue {
  "@odata.type": string;
  Id: number;
  Value: string;
}

export interface AuthorValue {
  "@odata.type": string;
  Claims: string;
  DisplayName: string;
  Email: string;
  Picture: string;
  Department: string;
  JobTitle: string;
}

export interface ReviewerValue {
  "@odata.type": string;
  Claims: string;
  DisplayName: string;
  Email: string;
  Picture: string;
  Department: string;
  JobTitle: string;
}

export interface Fulfiller1Value {
  "@odata.type": string;
  Claims: string;
  DisplayName: string;
  Email: string;
  Picture: string;
  Department: string;
  JobTitle: string;
}

export interface Fulfiller2Value {
  "@odata.type": string;
  Claims: string;
  DisplayName: string;
  Email: string;
  Picture: string;
  Department: string;
  JobTitle: string;
}

export interface Fulfiller3Value {
  "@odata.type": string;
  Claims: string;
  DisplayName: string;
  Email: string;
  Picture: string;
  Department: string;
  JobTitle: string;
}

export interface FulfillerActionedByValue {
  "@odata.type": string;
  Claims: string;
  DisplayName: string;
  Email: string;
  Picture: string;
  Department: string;
  JobTitle: string;
}

export interface StatusValue {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface AttachmentTypeValue {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface EditorValue {
  "@odata.type": string;
  Claims: string;
  DisplayName: string;
  Email: string;
  Picture: string;
  Department: string;
  JobTitle: string;
}

export interface IsFolderCustomValue {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface ApprovalSubmissionsBase {
  ID?: number;
  Title?: string;
  "NSC#Id"?: number;
  "Function#Id"?: number;
  "Category#Id"?: number;
  "Subcategory#Id"?: number;
  ContractingPartyName?: string;
  "Author#Claims"?: string;
  Created?: string;
  ContractValidity?: string;
  Remarks?: string;
  "Reviewer#Claims"?: string;
  ReviewerActionedDate?: string;
  ReviewerComments?: string;
  "Fulfiller1#Claims"?: string;
  "Fulfiller2#Claims"?: string;
  "Fulfiller3#Claims"?: string;
  "FulfillerActionedBy#Claims"?: string;
  FulfillerActionedDate?: string;
  FulfillerComments?: string;
  "Status#Id"?: number;
  Link?: string;
  "AttachmentType#Id"?: number[];
  "AttachmentType@odata.type"?: string;
  Modified?: string;
  "Editor#Claims"?: string;
  "IsFolderCustom#Id"?: number;
  OData__ColorTag?: string;
  ComplianceAssetId?: string;
  "{Identifier}"?: string;
  "{IsFolder}"?: boolean;
  "{Thumbnail}"?: {
    "@odata.type"?: string;
    // URL to the small, highly compressed size thumbnail of the item, if available.
    Small?: string;
    // URL to the standard size thumbnail of the item, if available.
    Medium?: string;
    // URL to the largest size thumbnail of the item, if available.
    Large?: string;
  };
  "{Link}"?: string;
  "{Name}"?: string;
  "{FilenameWithExtension}"?: string;
  "{Path}"?: string;
  "{FullPath}"?: string;
  "{ModerationStatus}"?: string;
  "{ModerationComment}"?: string;
  "{ContentType}#Id"?: string;
  "{ContentType}"?: {
    "@odata.type"?: string;
    Id?: string;
    Name?: string;
  };
  "{HasAttachments}"?: boolean;
  "{Attachments}@odata.type"?: string;
  "{Attachments}"?: {
   "@odata.type"?: string;
   Id?: string;
   AbsoluteUri?: string;
   DisplayName?: string;
 }[];
  "{VersionNumber}"?: string;
  "{TriggerWindowStartToken}"?: string;
  "{TriggerWindowEndToken}"?: string;
}

export interface ApprovalSubmissionsRead extends ApprovalSubmissionsBase {
  NSC: NSCValue;
  Function?: FunctionValue;
  Category?: CategoryValue;
  Subcategory?: SubcategoryValue;
  Author?: AuthorValue;
  Reviewer?: ReviewerValue;
  Fulfiller1?: Fulfiller1Value;
  Fulfiller2?: Fulfiller2Value;
  Fulfiller3?: Fulfiller3Value;
  FulfillerActionedBy?: FulfillerActionedByValue;
  Status?: StatusValue;
  AttachmentType?: AttachmentTypeValue[];
  Editor?: EditorValue;
  IsFolderCustom?: IsFolderCustomValue;
}